Added mechanism for saving and loading movable positions on script rerun
Usage:
Initialize Geomatplot with a file path, e.g.
Geomatplot(gca, "project.ini");
Then the positions of movables are saved to this file every time they are moved (this should get a setting later to be able to change behavior).
If the same script is run again and the file already exist, it loads the positions of the movables, then when a movable with a saved position is created, the position is overwritten by the saved state